refacotr: change to have one single layout
