+++ pokeapi/src/Layout.css
@@ -1,0 +1,21 @@
+body {
+  display: grid;
+  height: 100vh;
+  grid-template-columns: 100vw;
+  grid-template-rows: 80px auto;
+  gap: 0px 0px;
+  grid-template-areas:
+    "header"
+    "content";
+}
+
+.header {
+  display: flex;
+  grid-area: header;
+  z-index: 500;
+}
+
+.content {
+  grid-area: content;
+  display: flex;
+}
