+++ pokeapi/src/Pages/PokemonGame.tsx
@@ -1,31 +1,7 @@
 import Game from "../Organisms/Game/Game";
-import Navbar from "../Organisms/Header/Navbar";
 import LanguageSection from "../Organisms/LanguageSection/LanguageSection";
-import GameLayout from "../templates/GameLayout";
-import type { Link } from "../Types/Link";
-
+import GameTemplate from "../templates/GameTemplate";
 
 export default function PokemonGame() {
-
-    const links: Link[] = [
-        {
-            name: "Games",
-            link: "games",
-        },
-        {
-            name: "Langs",
-            link: "langs",
-        },
-        {
-            name: "Pokedex",
-            link: "pokedex",
-        },
-    ];
-
-    return (
-        <GameLayout
-            header={<Navbar links={links} />}
-            secondHeader={<LanguageSection />}
-            content={<Game />} />
-    )
+  return <GameTemplate secondHeader={<LanguageSection />} content={<Game />} />;
 }
