+++ pokeapi/src/templates/GameTemplate.tsx
@@ -1,0 +1,17 @@
+import type { ReactElement } from "react";
+import "./GameTemplate.css";
+
+export default function GameTemplate({
+  secondHeader,
+  content,
+}: {
+  secondHeader: ReactElement;
+  content: ReactElement;
+}) {
+  return (
+    <div className="gameContainer">
+      <div className="gameSecondHeader">{secondHeader}</div>
+      <div className="gameContent">{content}</div>
+    </div>
+  );
+}
