--- pokeapi/src/templates/layout.tsx
@@ -1,20 +1,0 @@
-import type { ReactElement } from "react";
-import "./layout.css";
-
-function Layout({ header, secondHeader, content }: { header: ReactElement, secondHeader: ReactElement, content: ReactElement }) {
-  return (
-    <div className="container">
-      <header className="header">
-        {header}
-      </header>
-      <div className="secondHeader">
-        {secondHeader}
-      </div>
-      <div className="content">
-        {content}
-      </div>
-    </div>
-  );
-}
-
-export default Layout;
